Updated dependencies and updated failing test. Updated travisci build
