+++ .travis.yml
@@ -1,7 +1,15 @@
 language: java
-dist: trusty
+dist: bionic
+
 jdk:
   - oraclejdk8
+  - openjdk11
   - openjdk12
+  - openjdk14
+
+cache:
+  directories:
+    - .autoconf
+    - $HOME/.m2
 before_install:
    - "export M2_HOME=/usr/local/maven"
